# openapi.yaml
openapi: 3.0.3
info:
  title: LeadOS API
  version: 0.1.0
  description: >
    AI-Powered CRM & Social Lead Automation Platform. Multi-tenant. Most routes
    require a Bearer JWT and an `X-Org-Id` header identifying the organization
    (tenant isolation, BRD §20). Import this file directly into Postman
    (File → Import) or any OpenAPI client.
servers:
  - url: http://localhost:4000
    description: Local
  - url: https://leados-api.onrender.com
    description: Production (example)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgId:
      name: X-Org-Id
      in: header
      required: true
      schema: { type: string }
      description: Organization id (tenant scope, BRD §20)
  schemas:
...
    Lead:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        source: { type: string, enum: [INSTAGRAM, WHATSAPP, FACEBOOK, WEBSITE, MANUAL, REFERRAL, CAMPAIGN, IMPORT] }
        status: { type: string, enum: [NEW, CONTACTED, QUALIFIED, PROPOSAL_SENT, NEGOTIATION, WON, LOST] }
        score: { type: integer }
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/leads:
    get:
      tags: [Leads]
      summary: List leads (paginated, filterable)
      parameters:
        - { $ref: '#/components/parameters/OrgId' }
        - { name: status, in: query, schema: { type: string } }
        - { name: source, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
      responses: { '200': { description: OK } }
    post:
      tags: [Leads]
      summary: Create lead
      parameters: [{ $ref: '#/components/parameters/OrgId' }]
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/Lead' } } } }
      responses: { '201': { description: Created }, '402': { description: Plan limit } }
  /api/v1/leads/export.csv:
    get:
      tags: [Leads]
      summary: Export leads as CSV
      parameters: [{ $ref: '#/components/parameters/OrgId' }]
      responses: { '200': { description: CSV, content: { text/csv: {} } } }
  /api/v1/leads/import:
    post:
      tags: [Leads]
      summary: Bulk import leads
      parameters: [{ $ref: '#/components/parameters/OrgId' }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { rows: { type: array, items: { $ref: '#/components/schemas/Lead' } } } } } } }
      responses: { '201': { description: Imported } }
  /api/v1/leads/bulk:
    post:
      tags: [Leads]
      summary: Bulk action on many leads (SET_STATUS | ASSIGN | DELETE)
      parameters: [{ $ref: '#/components/parameters/OrgId' }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { ids: { type: array, items: { type: string } }, action: { type: string }, status: { type: string }, assignedUserId: { type: string } } } } } }
      responses: { '200': { description: '{ affected }' } }
...
  /api/v1/leads/{id}/notes:
    post:
      tags: [Leads]
      summary: Add a note to a lead's activity timeline
      parameters: [{ $ref: '#/components/parameters/OrgId' }, { name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { note: { type: string } } } } } }
      responses: { '201': { description: Logged } }
  /api/v1/leads/{id}:
    get: { tags: [Leads], summary: Get lead, parameters: [{ $ref: '#/components/parameters/OrgId' }, { name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK }, '404': { description: Not found } } }
    patch: { tags: [Leads], summary: Update lead, parameters: [{ $ref: '#/components/parameters/OrgId' }, { name: id, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
    delete: { tags: [Leads], summary: Delete lead, parameters: [{ $ref: '#/components/parameters/OrgId' }, { name: id, in: path, required: true, schema: { type: string } }], responses: { '204': { description: Deleted } } }
...
  /api/v1/tasks:
    get: { tags: [Tasks], summary: List tasks, parameters: [{ $ref: '#/components/parameters/OrgId' }], responses: { '200': { description: OK } } }
    post: { tags: [Tasks], summary: Create task (syncs to calendar if due date), parameters: [{ $ref: '#/components/parameters/OrgId' }], responses: { '201': { description: Created } } }
...